+++ petanque/test/basic_api.ml
@@ -27,7 +27,10 @@ let start ~token =
   let root, uri = prepare_paths () in
   let* () = Petanque.Shell.set_workspace ~token ~debug ~root in
   let* () = Petanque.Shell.set_workspace ~token ~debug ~root in
-  Agent.start ~token ~uri ~thm:"rev_snoc_cons" ()
+  (* Careful to call [build_doc] before we have set an environment! [pet] and
+     [pet-server] are careful to always set a default one *)
+  let* doc = Petanque.Shell.build_doc ~token ~uri in
+  Agent.start ~token ~doc ~thm:"rev_snoc_cons" ()
 
 let extract_st (st : _ Agent.Run_result.t) =
   match st with